Check the expected data count when possible
An attempt to try and catch issues when the wrong verification method is used. Added extra checks to try and confirm that the correct number of data bytes are provided. There are still going to be issues with commands that have variable data counts as we won't pick up if there are checksum or mod256 bytes tacked onto the end.